+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 - Fixed range constructors for empty and null arrays and sequences
 - Added GrayAlpha pixel format constants
 - Added ToString() override for PixFormat
+- Added PixFileFormat.Webp and PixWebpSaveParams
 
 ### 5.3.10
 - [IO] Added SymMap constructor overload